Add files via upload
This is a simple C++ program originally written by another GitHub user. Its purpose is to let User enter the hour and minute of today's sunset time and will report back to user how long before it will be dark. This is useful if a person is getting ready to drive home from work right now and wants to know if it will be dark by the time he gets home. I would like to add some error-catching and logging functions to the program. I would like to figure out a way for the program to automatically know what is my sunset time today based on some weather app API or something like that.